archive/experimental/pilotkit/orchestrator/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==1.10.13
pandas==2.1.4
numpy==1.26.4
sqlmodel==0.0.14
sqlalchemy==1.4.49
pyarrow==14.0.2
sse-starlette==1.6.5
python-multipart==0.0.9
jinja2==3.1.3
matplotlib==3.8.2
scipy==1.11.4
typer==0.9.0
pyyaml==6.0.1
textdistance==4.6.1
